spearman for dataset that have no matching data

+++ Sample.cs
@@ -12,8 +12,8 @@ namespace CorrelationCalculator
 	public class Sample
 	{		
 		public double X { get; set; }
-		public int RankX { get; set; }
+		public double RankX { get; set; }
 		public double Y { get; set; }
-		public int RankY { get; set; }
+		public double RankY { get; set; }
 	}
 }
